Add files via upload
This is a simple trivia game utilizing the database and API from https://opentdb.com/.
For simplicity, when changing the questions you should copy paste this code's available database then go to ChatGPT and copy paste it there along with the new database from open trivia database. Then, ask it to format it in a similar fashion and compare what you got.

Code was inspired from a Python Bootcamp course.